👌 IMPROVE: Docs and Generative UI code rework
