+++ src/lib/chat/external-apis.ts
@@ -1,0 +1,14 @@
+export async function getWeather() {
+  //Make an API Call here!!
+  return {
+    weather: {
+      monday: "30ºC",
+      tuesday: "28ºC",
+      wednesday: "32ºC",
+      thursday: "29ºC",
+      friday: "27ºC",
+      saturday: "28ºC",
+      sunday: "29ºC",
+    },
+  };
+}
